% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform_deps.R
\name{check_wsl}
\alias{check_wsl}
\title{Check if WSL is available and configured}
\usage{
check_wsl(distro = "Ubuntu-22.04")
}
\arguments{
\item{distro}{WSL distribution name}
}
\value{
Logical indicating success
}
\description{
Check if WSL is available and configured
}
