% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{clean_transcript_id}
\alias{clean_transcript_id}
\title{Clean a transcript/gene ID for exact-match purposes}
\usage{
clean_transcript_id(x)
}
\description{
Handles:
  ENST00000456328.2
  ENST00000456328.2|ENSG00000223972.5|DDX11L1|...
  ENST00000456328.2 some description
  ENSG00000223972.5
}
\keyword{internal}
