% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dge.R
\name{create_dds_object}
\alias{create_dds_object}
\title{Create DESeqDataSet object}
\usage{
create_dds_object(tx_data, level, base, model, replicate_col)
}
\arguments{
\item{tx_data}{List returned by import_counts}

\item{level}{Target level}

\item{base}{Reference level}

\item{model}{Design formula as a string}

\item{replicate_col}{Column containing replicate information}
}
\value{
DESeqDataSet object
}
\description{
Create DESeqDataSet object
}
