% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.expressom_rmd_path}
\alias{.expressom_rmd_path}
\title{Locate a bundled template/Rmd file under inst/rmd/}
\usage{
.expressom_rmd_path(filename)
}
\description{
More robust than the original version:
- checks installed package first
- checks common source-checkout locations
- walks up from the working directory
- honors EXPRESSOM_RMD_DIR
}
\keyword{internal}
