% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.fill_entrez_with_bitr}
\alias{.fill_entrez_with_bitr}
\title{Fill missing Entrez IDs in a gene_map using clusterProfiler::bitr}
\usage{
.fill_entrez_with_bitr(
  gene_map,
  org_obj,
  id_col = "ensembl",
  symbol_col = "symbol"
)
}
\description{
Fill missing Entrez IDs in a gene_map using clusterProfiler::bitr
}
\keyword{internal}
