% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform_deps.R
\name{.find_cpat_hexamer}
\alias{.find_cpat_hexamer}
\title{Locate the CPAT hexamer frequency table for a given organism}
\usage{
.find_cpat_hexamer(
  organism = "Human",
  wsl_distro = "Ubuntu-22.04",
  use_wsl = TRUE,
  conda_sh = NULL,
  conda_env = "isoform_tools",
  log_dir = NULL
)
}
\description{
Locate the CPAT hexamer frequency table for a given organism
}
\keyword{internal}
