% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform.R
\name{.find_cpat_logit_model}
\alias{.find_cpat_logit_model}
\title{Locate the CPAT logit model RData file for a given organism}
\usage{
.find_cpat_logit_model(
  organism = "Human",
  wsl_distro = "Ubuntu",
  use_wsl = TRUE,
  conda_sh = NULL,
  conda_env = "isoform_tools",
  log_dir = NULL
)
}
\description{
Locate the CPAT logit model RData file for a given organism
}
\keyword{internal}
