% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.load_org_db}
\alias{.load_org_db}
\title{Load an OrgDb annotation package by name (e.g. "org.Hs.eg.db")}
\usage{
.load_org_db(org_db_name)
}
\description{
Load an OrgDb annotation package by name (e.g. "org.Hs.eg.db")
}
\keyword{internal}
