% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform_predictors.R
\name{.pfam_from_hmmscan_domtblout}
\alias{.pfam_from_hmmscan_domtblout}
\title{Convert hmmscan '--domtblout' output into pfam_scan.pl format}
\usage{
.pfam_from_hmmscan_domtblout(domtblout_path, dest_path)
}
\description{
Convert hmmscan '--domtblout' output into pfam_scan.pl format
}
\keyword{internal}
