% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform_predictors.R
\name{.pfam_from_interproscan_tsv}
\alias{.pfam_from_interproscan_tsv}
\title{Convert an InterProScan '-f tsv' (Pfam-filtered) file into pfam_scan.pl format}
\usage{
.pfam_from_interproscan_tsv(ips_tsv_path, dest_path)
}
\description{
Convert an InterProScan '-f tsv' (Pfam-filtered) file into pfam_scan.pl format
}
\keyword{internal}
