% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{.render_placeholder_template}
\alias{.render_placeholder_template}
\title{Render a {{PLACEHOLDER}}-style template to a temp file with values substituted}
\usage{
.render_placeholder_template(template_file, values, fileext = ".Rmd")
}
\description{
Render a {{PLACEHOLDER}}-style template to a temp file with values substituted
}
\keyword{internal}
