% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{.report_unmatched_labels}
\alias{.report_unmatched_labels}
\title{Tell the user which relabelling entries / samples did not line up}
\usage{
.report_unmatched_labels(sample_labels, sample_ids)
}
\description{
Tell the user which relabelling entries / samples did not line up
}
\keyword{internal}
