% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_functional.R
\name{.run_go_gsea}
\alias{.run_go_gsea}
\title{Run GSEA for Gene Ontology (BP, MF, CC)}
\usage{
.run_go_gsea(gene_list, org_db, ont, pvalue_cutoff, out_dir, comp_name)
}
\description{
Run GSEA for Gene Ontology (BP, MF, CC)
}
\keyword{internal}
