% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_functional.R
\name{.run_kegg_pathview}
\alias{.run_kegg_pathview}
\title{Run KEGG Pathview for GSEA results}
\usage{
.run_kegg_pathview(gseaKEGG, expression_vector, kegg_code, top_genes, dir_kegg)
}
\description{
Run KEGG Pathview for GSEA results
}
\keyword{internal}
