% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wsl_core.R
\name{.to_wsl_path}
\alias{.to_wsl_path}
\title{Convert a Windows file path to a WSL-compatible Unix path}
\usage{
.to_wsl_path(win_path, distro = "Ubuntu-22.04")
}
\description{
Convert a Windows file path to a WSL-compatible Unix path
}
\keyword{internal}
