% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{download_ensembl_refs}
\alias{download_ensembl_refs}
\title{Download Ensembl Reference FASTA and GTF}
\usage{
download_ensembl_refs(ensembl_package_name, out_dir = "./reference")
}
\description{
Download Ensembl Reference FASTA and GTF
}
