% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform.R
\name{import_transcript_counts}
\alias{import_transcript_counts}
\title{Import transcript-level counts for isoform analysis}
\usage{
import_transcript_counts(
  data_dir,
  sample_table,
  ensembl_package_name,
  count_type = "salmon",
  matrix_file = NULL,
  subset_sample = NULL,
  remove_sample = NULL,
  custom_tx2gene = NULL,
  custom_gene_map = NULL
)
}
\description{
Import transcript-level counts for isoform analysis
}
