% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{plot_l2fc_heatmap}
\alias{plot_l2fc_heatmap}
\title{Plot Log2 Fold Change Heatmap (1-Column Format)}
\usage{
plot_l2fc_heatmap(dds, selected_genes, condition_col, level, base, plot_dir)
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{selected_genes}{Character vector of gene symbols to plot}

\item{condition_col}{Character string representing the design metadata column}

\item{level}{String representing the foreground group (e.g. KO)}

\item{base}{String representing the background reference group (e.g. WT)}

\item{plot_dir}{String directory path where the PDF will be saved}
}
\description{
Plot Log2 Fold Change Heatmap (1-Column Format)
}
