% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{plot_sample_correlation}
\alias{plot_sample_correlation}
\title{Sample Correlation Heatmap}
\usage{
plot_sample_correlation(vsd, condition_col, plot_dir, comp_name)
}
\arguments{
\item{vsd}{VST-transformed DESeqDataSet}

\item{condition_col}{Column name for sample annotation}

\item{plot_dir}{Directory to save PDF}

\item{comp_name}{Comparison name for file naming}
}
\description{
Sample Correlation Heatmap
}
