% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{plot_sample_zscore}
\alias{plot_sample_zscore}
\title{Plot Individual Sample Z-Score Heatmap}
\usage{
plot_sample_zscore(dds, selected_genes, condition_col, level, base, plot_dir)
}
\arguments{
\item{dds}{A DESeqDataSet object}

\item{selected_genes}{Character vector of gene symbols to plot}

\item{condition_col}{Character string representing the design metadata column}

\item{level}{String representing the foreground group}

\item{base}{String representing the background reference group}

\item{plot_dir}{String directory path where the PDF will be saved}
}
\description{
Plot Individual Sample Z-Score Heatmap
}
