% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{plotP_fork}
\alias{plotP_fork}
\title{Fork to SPIA Plot function (ggplot2 + ggrepel upgrade)}
\usage{
plotP_fork(x, threshold = 0.01)
}
\arguments{
\item{x}{A SPIA results data.frame}

\item{threshold}{Threshold for significance}
}
\value{
Generates a plot
}
\description{
Fork to SPIA Plot function (ggplot2 + ggrepel upgrade)
}
