% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform_dte.R
\name{run_isoform_pca}
\alias{run_isoform_pca}
\title{Transcript-level PCA}
\usage{
run_isoform_pca(
  isoform_obj,
  condition,
  level = NULL,
  base = NULL,
  out_dir,
  batch_col = NULL,
  pca_ntop = 500,
  save_dir = NULL,
  pca_colors = NULL,
  sample_labels = NULL
)
}
\arguments{
\item{pca_colors, sample_labels}{See `plot_custom_pca()`}
}
\description{
Transcript-level PCA
}
