% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{validate_environment}
\alias{validate_environment}
\title{ExpressOM Pre-Flight Environment Validation}
\usage{
validate_environment(run_isoform = TRUE, run_functional = TRUE)
}
\description{
ExpressOM Pre-Flight Environment Validation
}
