% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform_utils.R
\name{convert_pdf_to_png}
\alias{convert_pdf_to_png}
\title{Convert a PDF plot to a PNG alongside it}
\usage{
convert_pdf_to_png(pdf_file, dpi = 200)
}
\description{
Convert a PDF plot to a PNG alongside it
}
\keyword{internal}
