% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_processing.R
\name{create_dds_object}
\alias{create_dds_object}
\title{Create DESeqDataSet object}
\usage{
create_dds_object(txi, meta, level, base, model, replicate_col)
}
\arguments{
\item{txi}{tximport object}

\item{meta}{Metadata dataframe}

\item{level}{Target level}

\item{base}{Reference level}

\item{model}{Design formula as a string}

\item{replicate_col}{Column containing replicate information}
}
\value{
DESeqDataSet object
}
\description{
Create DESeqDataSet object
}
