% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{create_homemade_db}
\alias{create_homemade_db}
\title{Create and Bundle Homemade Ensembl Database}
\usage{
create_homemade_db(
  species = "human",
  release = "107",
  maintainer = "User <user@example.com>",
  author = "ExpressOM Builder"
)
}
\description{
Create and Bundle Homemade Ensembl Database
}
