% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{create_homemade_db}
\alias{create_homemade_db}
\title{Create and Bundle Homemade Ensembl Database}
\usage{
create_homemade_db(species = "human", release = "107")
}
\arguments{
\item{species}{String: "human" or "mouse"}

\item{release}{Ensembl release version (e.g., "107")}
}
\description{
Create and Bundle Homemade Ensembl Database
}
