% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform.R
\name{debug_wsl}
\alias{debug_wsl}
\title{Debug the external-predictor execution environment}
\usage{
debug_wsl(
  distro = "Ubuntu",
  out_dir = NULL,
  log_dir = NULL,
  conda_env = "isoform_tools",
  verbose = TRUE,
  use_wsl = NULL
)
}
\description{
Debug the external-predictor execution environment
}
