% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wsl_core.R
\name{.find_conda_sh}
\alias{.find_conda_sh}
\title{Best-effort discovery of a conda profile script (conda.sh)}
\usage{
.find_conda_sh(wsl_distro = "Ubuntu-22.04", use_wsl = TRUE, log_dir = NULL)
}
\description{
Best-effort discovery of a conda profile script (conda.sh)
}
\keyword{internal}
