% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform.R
\name{.find_pfam_db}
\alias{.find_pfam_db}
\title{Find the Pfam-A.hmm database in common paths inside the execution environment}
\usage{
.find_pfam_db(
  wsl_distro = "Ubuntu",
  use_wsl = TRUE,
  conda_sh = NULL,
  conda_env = "isoform_tools",
  log_dir = NULL
)
}
\description{
Find the Pfam-A.hmm database in common paths inside the execution environment
}
\keyword{internal}
