% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wsl_core.R
\name{.log_wsl_command}
\alias{.log_wsl_command}
\title{Write a WSL command execution to log}
\usage{
.log_wsl_command(cmd, exit_code, stdout = NULL, stderr = NULL, log_dir)
}
\description{
Write a WSL command execution to log
}
\keyword{internal}
