% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dge.R
\name{.write_dge_log}
\alias{.write_dge_log}
\title{Write structured DGE parameters log}
\usage{
.write_dge_log(
  out_dir,
  comp_name,
  model,
  test,
  reduced,
  level,
  base,
  shrink_method,
  padj_cutoff,
  n_genes_input,
  n_genes_after_filter,
  n_de_genes_up,
  n_de_genes_down,
  filter_criterion = "rowSums(counts) >= 1"
)
}
\description{
Write structured DGE parameters log
}
\keyword{internal}
