% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform.R
\name{.wsl_tool_exists}
\alias{.wsl_tool_exists}
\title{Check whether a command-line tool is accessible in the execution environment}
\usage{
.wsl_tool_exists(
  tool_name,
  wsl_distro = "Ubuntu",
  use_wsl = TRUE,
  conda_sh = NULL,
  conda_env = "isoform_tools"
)
}
\description{
Check whether a command-line tool is accessible in the execution environment
}
\keyword{internal}
