% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{download_ensembl_refs}
\alias{download_ensembl_refs}
\title{Download Ensembl Reference FASTA and GTF}
\usage{
download_ensembl_refs(ensembl_package_name, out_dir = "./reference")
}
\arguments{
\item{ensembl_package_name}{Name of the Ensembl package (e.g., "EnsDb.Hsapiens.v107")}

\item{out_dir}{Directory where the reference files should be saved}
}
\value{
A list containing the local paths: `gtf`, `cdna_fasta`, `ncrna_fasta`.
}
\description{
Download Ensembl Reference FASTA and GTF
}
