% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_processing.R
\name{export_significant_results}
\alias{export_significant_results}
\title{Export Significant Results}
\usage{
export_significant_results(
  res_shrunken,
  res_unshrunken,
  dds,
  out_dir,
  level,
  base,
  gene_map,
  padj_cutoff
)
}
\arguments{
\item{res_shrunken}{Shrunken DESeq results}

\item{res_unshrunken}{Unshrunken DESeq results}

\item{dds}{DESeqDataSet object}

\item{out_dir}{Output directory}

\item{level}{Comparison level}

\item{base}{Base level}

\item{gene_map}{Gene symbol mapping dataframe}

\item{padj_cutoff}{Adjusted p-value cutoff}
}
\value{
List of processed dataframes
}
\description{
Export Significant Results
}
