% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_visualizations.R
\name{generate_bulk_visualizations}
\alias{generate_bulk_visualizations}
\title{Generate Bulk Visualizations}
\usage{
generate_bulk_visualizations(
  dds,
  edb,
  res_shrunken,
  res_unshrunken,
  results_data,
  out_dir,
  level,
  base,
  main_condition,
  top_genes,
  padj_cutoff
)
}
\arguments{
\item{dds}{DESeq object}

\item{edb}{Ensembl database object}

\item{res_shrunken}{Shrunken DE results}

\item{res_unshrunken}{Unshrunken DE results}

\item{results_data}{Object from export_significant_results}

\item{out_dir}{Output directory}

\item{level}{Level}

\item{base}{Base}

\item{main_condition}{Extracted primary condition factor}

\item{top_genes}{N top genes}

\item{padj_cutoff}{Adjusted p-value significance cutoff}
}
\description{
This function generates a series of visualizations to summarize the differential expression results.
}
