% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{get_organism_info}
\alias{get_organism_info}
\title{Extract organism specific databases and properties}
\usage{
get_organism_info(edb)
}
\description{
Extract organism specific databases and properties
}
