% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{get_organism_info}
\alias{get_organism_info}
\title{Extract organism specific databases and properties}
\usage{
get_organism_info(edb)
}
\arguments{
\item{edb}{Ensembl database object}
}
\description{
Extract organism specific databases and properties
}
