% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dge.R
\name{import_counts}
\alias{import_counts}
\title{Import RNA-seq Counts and Prepare Metadata}
\usage{
import_counts(
  data_dir,
  sample_table,
  ensembl_package_name,
  count_type = "salmon",
  out_dir,
  matrix_file = NULL,
  subset_sample = NULL,
  remove_sample = NULL
)
}
\arguments{
\item{data_dir}{Folder where the input data is stored.}

\item{sample_table}{Path to the sample table.}

\item{ensembl_package_name}{Name of the installed Ensembl database package.}

\item{count_type}{Type of RNA count (e.g., "salmon" or "matrix").}

\item{out_dir}{Output directory.}

\item{matrix_file}{Path to raw counts file if count_type is "matrix".}

\item{subset_sample}{Optional string to filter the sample table.}

\item{remove_sample}{Optional character vector of sample IDs to exclude}
}
\value{
A list containing txi (or counts), meta, edb, gene_map, and type.
}
\description{
Import RNA-seq Counts and Prepare Metadata
}
