% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_processing.R
\name{import_counts}
\alias{import_counts}
\title{Import RNA-seq Counts and Prepare Metadata}
\usage{
import_counts(
  data_dir,
  sample_table,
  ensembl_package_name,
  count_type = "salmon",
  out_dir
)
}
\arguments{
\item{data_dir}{Folder where the input data is stored.}

\item{sample_table}{Path to the sample table.}

\item{ensembl_package_name}{Name of the installed Ensembl database package.}

\item{count_type}{Type of RNA count (e.g., "salmon").}

\item{out_dir}{Output directory.}
}
\value{
A list containing txi, meta, edb, and a gene_symbol_map.
}
\description{
Import RNA-seq Counts and Prepare Metadata
}
