% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform.R
\name{import_transcript_counts}
\alias{import_transcript_counts}
\title{Import transcript-level counts for isoform analysis}
\usage{
import_transcript_counts(
  data_dir,
  sample_table,
  ensembl_package_name,
  count_type = "salmon",
  matrix_file = NULL,
  subset_sample = NULL,
  remove_sample = NULL
)
}
\arguments{
\item{data_dir}{Folder with quantification files}

\item{sample_table}{Path to sample metadata}

\item{ensembl_package_name}{EnsDb package name}

\item{count_type}{Quantification type ("salmon", "kallisto", etc.)}

\item{matrix_file}{Path to raw counts matrix (if count_type = "matrix")}

\item{subset_sample}{Optional filter expression}

\item{remove_sample}{Optional sample IDs to exclude}
}
\value{
List with txi (transcript counts), meta, tx2gene, gene_map
}
\description{
Import transcript-level counts for isoform analysis
}
