% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{install_internal_db}
\alias{install_internal_db}
\title{Install Bundled Ensembl Database}
\usage{
install_internal_db(pkg_name = NULL)
}
\arguments{
\item{pkg_name}{Optional string to match a specific database (e.g., "Mmusculus" or "Hsapiens")}
}
\description{
Install Bundled Ensembl Database
}
