% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform.R
\name{install_isoform_databases}
\alias{install_isoform_databases}
\title{Install required databases for CPAT, Pfam, and SignalP}
\usage{
install_isoform_databases(
  distro = "Ubuntu",
  use_wsl = (.Platform$OS.type == "windows"),
  cpat_data_dir = NULL,
  pfam_db_dir = NULL,
  conda_env = "isoform_tools",
  log_dir = NULL
)
}
\description{
Install required databases for CPAT, Pfam, and SignalP
}
