% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform.R
\name{install_signalp_from_windows}
\alias{install_signalp_from_windows}
\title{Install SignalP from a Windows directory into WSL}
\usage{
install_signalp_from_windows(
  windows_signalp_dir,
  distro = "Ubuntu",
  install_path = "/usr/local/signalp",
  log_dir = NULL
)
}
\description{
Install SignalP from a Windows directory into WSL
}
