% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform.R
\name{install_wsl_isoform_tools}
\alias{install_wsl_isoform_tools}
\title{Install required external tools inside WSL using mamba or apt/pip}
\usage{
install_wsl_isoform_tools(
  distro = "Ubuntu",
  use_mamba = TRUE,
  install_databases = TRUE,
  windows_signalp_dir = NULL,
  log_dir = NULL
)
}
\description{
Install required external tools inside WSL using mamba or apt/pip
}
