% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_isoform.R
\name{load_isoform_results}
\alias{load_isoform_results}
\title{Load previously saved isoform analysis step checkpoints}
\usage{
load_isoform_results(save_dir)
}
\description{
Load previously saved isoform analysis step checkpoints
}
