% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform_dtu.R
\name{plot_dexseq_gene}
\alias{plot_dexseq_gene}
\title{DEXSeq-style transcript usage plot for a single gene}
\usage{
plot_dexseq_gene(
  dxr_list,
  gene_id,
  plot_dir,
  gene_symbol = NULL,
  splicing = FALSE
)
}
\description{
DEXSeq-style transcript usage plot for a single gene
}
