% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spia_utils.R
\name{plotP_fork}
\alias{plotP_fork}
\title{Fork to SPIA Plot function}
\usage{
plotP_fork(x, threshold = 0.01)
}
\arguments{
\item{x}{A SPIA results data.frame}

\item{threshold}{Threshold for significance}
}
\value{
NULL (generates a plot)
}
\description{
This function generates a two-way evidence plot for SPIA results, showing the relationship between the p-values for NDE (Number of Differentially Expressed genes) and PERT (Perturbation). It also highlights significant pathways based on the specified threshold and indicates the combined p-value method used (Fisher or Normal Inverse). The plot includes lines representing the significance thresholds and labels for significant pathways.
}
