% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform.R
\name{run_dtu}
\alias{run_dtu}
\title{Run DTU using DRIMSeq}
\usage{
run_dtu(
  isoform_obj,
  condition,
  level,
  base,
  min_gene_expr = 10,
  min_transcript_expr = 0.05,
  min_samps_gene_expr = NULL,
  min_samps_feature_expr = 3,
  chunk_size = 5000,
  max_transcripts = 300,
  min_transcript_total = 10,
  bpparam = NULL
)
}
\description{
Run DTU using DRIMSeq
}
