% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_visualizations.R
\name{run_eda}
\alias{run_eda}
\title{Exploratory Data Analysis for DESeq2}
\usage{
run_eda(dds, edb, out_dir, level, base, main_condition)
}
\arguments{
\item{dds}{DESeqDataSet object}

\item{edb}{Ensembl Database}

\item{out_dir}{Output directory}

\item{level}{Level comparison (e.g. "Treated")}

\item{base}{Base Level (e.g. "Control")}

\item{main_condition}{The primary modeled condition}
}
\description{
Exploratory Data Analysis for DESeq2
}
