% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_functional.R
\name{run_functional_analysis}
\alias{run_functional_analysis}
\title{Run Functional Analysis with Directional Stat Management}
\usage{
run_functional_analysis(
  res_tbl,
  sig_res,
  edb,
  out_dir,
  level,
  base,
  top_genes,
  padj_cutoff = 0.01,
  go_pvalue_cutoff = 0.05,
  go_qvalue_cutoff = 0.2,
  gsea_metric = "stat",
  test_type = "Wald",
  run_spia = FALSE
)
}
\arguments{
\item{res_tbl}{Results table}

\item{sig_res}{Significant list}

\item{edb}{Ensembl database object}

\item{out_dir}{Output dir}

\item{level}{Target condition}

\item{base}{Base condition}

\item{top_genes}{Limit for GO dotplots}

\item{padj_cutoff}{Adjusted p-value significance cutoff}

\item{go_pvalue_cutoff}{GO ORA raw p-value cutoff (default 0.05)}

\item{go_qvalue_cutoff}{GO ORA q-value cutoff (default 0.2)}

\item{gsea_metric}{Metric to rank genes for GSEA ("stat", "signed_pval", or "log2FoldChange")}

\item{test_type}{The upstream test design used: "Wald" or "LRT" (Required for correct stat ranking)}
}
\value{
List with functional results
}
\description{
Run Functional Analysis with Directional Stat Management
}
