% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_functional.R
\name{run_functional_analysis}
\alias{run_functional_analysis}
\title{Run Functional Analysis}
\usage{
run_functional_analysis(
  res_tbl,
  sig_res,
  edb,
  out_dir,
  level,
  base,
  top_genes,
  padj_cutoff
)
}
\arguments{
\item{res_tbl}{Results table}

\item{sig_res}{Significant list}

\item{edb}{Ensembl database object}

\item{out_dir}{Output dir}

\item{level}{Target condition}

\item{base}{Base condition}

\item{top_genes}{Limit for GO dotplots}

\item{padj_cutoff}{Adjusted p-value significance cutoff}
}
\value{
List with functional results
}
\description{
Run Functional Analysis
}
