% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform.R
\name{run_isoform_pca}
\alias{run_isoform_pca}
\title{Transcript-level PCA}
\usage{
run_isoform_pca(
  isoform_obj,
  condition,
  level = NULL,
  base = NULL,
  out_dir,
  batch_col = NULL,
  pca_ntop = 500
)
}
\description{
Transcript-level PCA
}
