% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_isoform_switch.R
\name{run_isoform_switch}
\alias{run_isoform_switch}
\title{Run IsoformSwitchAnalyzeR analysis}
\usage{
run_isoform_switch(
  dte_results = NULL,
  dtu_results = NULL,
  isoform_obj,
  condition,
  level,
  base,
  fasta_file,
  gff_file,
  out_dir,
  run_predictors = FALSE,
  use_wsl = (.Platform$OS.type == "windows"),
  wsl_distro = "Ubuntu-22.04",
  save_dir = NULL,
  resume_from = NULL,
  bsgenome_name = NULL,
  predictor_cpu = NULL,
  log_dir = NULL,
  custom_transcript_id_map = NULL,
  skip_fasta_filter = FALSE,
  test_engine = c("DEXSeq", "DRIMSeq", "satuRn"),
  organism = NULL
)
}
\description{
Run IsoformSwitchAnalyzeR analysis
}
