% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{validate_environment}
\alias{validate_environment}
\title{ExpressOM Pre-Flight Environment Validation}
\usage{
validate_environment(run_isoform = TRUE, run_functional = TRUE)
}
\arguments{
\item{run_isoform}{Logical; checks isoform dependencies if TRUE.}

\item{run_functional}{Logical; checks enrichment dependencies if TRUE.}
}
\value{
Throws an error if dependencies are missing, returns TRUE silently otherwise.
}
\description{
ExpressOM Pre-Flight Environment Validation
}
